app/requirements.txt
mysql-connector
Flask
prometheus-flask-exporter
